+++ tasks/gulp/asset-version.js
@@ -20,14 +20,13 @@ gulp.task('update-assets-version', () => {
   let pkg = require('../../' + configPaths.packages + 'all/package.json')
   fs.writeFileSync(taskArguments.destination + '/VERSION.txt', pkg.version + '\r\n')
   return gulp.src([
-    taskArguments.destination + '/css/govuk-frontend.min.css',
-    taskArguments.destination + '/css/govuk-frontend-old-ie.min.css',
-    taskArguments.destination + '/js/govuk-frontend.min.js'
+    taskArguments.destination + '/govuk-frontend.min.css',
+    taskArguments.destination + '/govuk-frontend-old-ie.min.css',
+    taskArguments.destination + '/govuk-frontend.min.js'
   ])
   .pipe(vinylPaths(del))
   .pipe(gulpif(isDist,
     rename(obj => {
-      obj.dirname += '/' + obj.extname.replace('.', '')
       obj.basename = obj.basename.replace(/(govuk.*)(?=\.min)/g, '$1-' + pkg.version)
       return obj
     })