+++ tasks/gulp/compile-assets.js
@@ -56,7 +56,13 @@ gulp.task('scss:compile', () => {
         extname: '.min.css'
       })
     ))
-    .pipe(gulp.dest(taskArguments.destination + '/css/'))
+    .pipe(gulp.dest(
+      gulpif(
+        isDist,
+        taskArguments.destination + '/',
+        taskArguments.destination + '/css/'
+      )
+    ))
 
   let compileOldIe = gulp.src(compileOldIeStyleshet)
     .pipe(plumber(errorHandler))
@@ -90,7 +96,13 @@ gulp.task('scss:compile', () => {
         extname: '.min.css'
       })
     ))
-    .pipe(gulp.dest(taskArguments.destination + '/css/'))
+    .pipe(gulp.dest(
+      gulpif(
+        isDist,
+        taskArguments.destination + '/',
+        taskArguments.destination + '/css/'
+      )
+    ))
 
   return merge(compile, compileOldIe)
 })
@@ -98,9 +110,11 @@ gulp.task('scss:compile', () => {
 // Compile js task for preview ----------
 // --------------------------------------
 gulp.task('js:compile', () => {
+  // for dist/ folder we only want compiled 'all.js' file
+  let srcFiles = isDist ? configPaths.src + 'all/*.js' : configPaths.src + '**/*.js'
   return gulp.src([
     '!' + configPaths.src + '**/*.test.js',
-    configPaths.src + '**/*.js'
+    srcFiles
   ])
     .pipe(rollup({
       // Legacy mode is required for IE8 support
@@ -111,16 +125,10 @@ gulp.task('js:compile', () => {
     .pipe(gulpif(isDist, uglify()))
     .pipe(gulpif(isDist,
       rename({
+        basename: 'govuk-frontend',
         extname: '.min.js'
       })
     ))
     .pipe(eol())
-    .pipe(gulp.dest(
-      // output files to dist/components if destination is dist, otherwise copy to packages/
-      gulpif(
-        isDist,
-        taskArguments.destination + '/components/',
-        taskArguments.destination + '/'
-      ))
-    )
+    .pipe(gulp.dest(taskArguments.destination + '/'))
 })